back-end/requirements.txt
# Core web framework
flask
flask-cors

# HTTP requests
requests

# AI/ML
openai

# Vector database (for RAG)
chromadb

# Text splitting
langchain
langchain-community

# PDF reading (for ingesting literature)
pypdf

# Environment variables
python-dotenv

# Database
pymongo>=4.0
dnspython>=2.0

# Production server
gunicorn==21.2.0
